refactor: rename create-feature-harness-options
